+++ CHANGELOG.md
@@ -59,3 +59,4 @@
 * Fixed an SVG width issue where if no data was supplied, the width and height were set to -15, which threw an exception.
 * Fixed a bug where if the tooltip width and height got too big, the arrow wasn't pointing at the child node.
 * Fixed a bug where the width of the SVG was being determined incorrectly.
+* Added a way to not show the value on the tooltip by setting the `valueKeyName` to an empty string.